modified Makefile
