+++ .vscode/settings.json
@@ -1,5 +1,8 @@
 {
     "files.associations": {
-        "lexer.h": "c"
+        "lexer.h": "c",
+        "struct.h": "c",
+        "minishell.h": "c",
+        "parser.h": "c"
     }
 }
