+++ src/parser.c
@@ -3,21 +3,22 @@
 /*                                                        :::      ::::::::   */
 /*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: koseki.yusuke <koseki.yusuke@student.42    +#+  +:+       +#+        */
+/*   By: sakitaha <sakitaha@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/21 22:51:59 by koseki.yusu       #+#    #+#             */
-/*   Updated: 2024/03/21 23:55:01 by koseki.yusu      ###   ########.fr       */
+/*   Updated: 2024/04/12 02:13:16 by sakitaha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 
-#include "../include/lexer.h"
-#include "../include/parser.h"
-#include "../include/minishell.h"
+#include "lexer.h"
+#include "parser.h"
+#include "minishell.h"
 
-t_ast *parse_token(mgr)
+t_ast *parse_token(t_mgr *mgr)
 {
-    
+
 
     // ND_PIPEなら新しいnode作って右につける
+    return 0;
 }
